Extract OIDC UI
