+++ index.html
@@ -109,9 +109,6 @@
             <dialog id="uploadDialog" is="solid-upload-dialog"></dialog>
             <dialog id="notFoundDialog" is="solid-ok-dialog" data-title="Not found"></dialog>
             <dialog id="newContainerDialog" is="solid-input-dialog" data-title="New container" data-label="Container name" data-required="true"></dialog>
-
-            <dialog id="storageSelectDialog" is="solid-select-dialog" data-title="Select storage URI" data-label="Storage URIs"></dialog>
-            <dialog id="idpSelectDialog" is="solid-select-dialog" data-title="Choose IDP" data-label="IDPs"></dialog>
         </div>
         <menu style="grid-area: menu">
             <li>
