+++ scripts/elements/index.js
@@ -7,7 +7,6 @@ import {ContainerViewItemElement} from "./ContainerViewItemElement.js"
 import {ResourceViewerElement} from "./ResourceViewerElement.js"
 import {AddressBarElement} from "./AddressBarElement.js"
 import {CrumbTrailElement} from "./CrumbTrailElement.js"
-import {AuthenticationDialog} from "./AuthenticationDialog.js"
 import {UploadDialog} from "./UploadDialog.js"
 import {GrantRequestDialog} from "./GrantRequestDialog.js"
 import {GrantResponseDialog} from "./GrantResponseDialog.js"
@@ -19,7 +18,6 @@ customElements.define("solid-container-view-item", ContainerViewItemElement, {ex
 customElements.define("solid-resource-viewer", ResourceViewerElement, {extends: "iframe"})
 customElements.define("solid-address-bar", AddressBarElement, {extends: "form"})
 customElements.define("solid-crumb-trail", CrumbTrailElement, {extends: "ul"})
-customElements.define("solid-authentication-dialog", AuthenticationDialog, {extends: "dialog"})
 customElements.define("solid-upload-dialog", UploadDialog, {extends: "dialog"})
 customElements.define("solid-grant-request-dialog", GrantRequestDialog, {extends: "dialog"})
 customElements.define("solid-grant-response-dialog", GrantResponseDialog, {extends: "dialog"})
